% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess.R
\name{createSCEobject}
\alias{createSCEobject}
\title{Returns a list with a SingleCellExperiment object}
\usage{
createSCEobject(
  xx,
  assay.name.raw = "counts",
  assay.name.normalization = "logcounts",
  metadata = NULL,
  toFactors = NULL,
  chosen.hvg = NULL,
  nHVGs = 3000,
  nPCs = 50,
  calcRedDim = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{xx}{Either a matrix with counts, or a SCE or Seurat object.}

\item{assay.name.raw}{Assay name for raw counts matrix if object is a SCE. Defaults to \code{counts}.}

\item{assay.name.normalization}{Assay name for normalized matrix if present in SCE. If not present,
it calculates \code{logcounts} (default).}

\item{metadata}{Optional dataframe with cell metadata.}

\item{toFactors}{Optional metadata column names (or colData) to use as factors. If \code{NULL} (default),
a quick clusterization will be computed.}

\item{chosen.hvg}{Optional list of Highly Variable Genes.}

\item{nHVGs}{Number of Highly Variable Genes to use if \code{chosen.hvg=NULL}. Defaults to 3000.}

\item{nPCs}{Number of Principal Components to use in PCA. Defaults to 50.}

\item{calcRedDim}{Whether to compute reduced dimensions (PCA, UMAP, TSNE, UMAP2D, TSNE2D) or not.
Defaults to \code{TRUE}.}

\item{verbose}{Step by step status while function is running. Defaults to \code{TRUE}.}
}
\value{
List with a SingleCellExperiment object and additional data ready for use in scXplorer.
}
\description{
\code{createSCEobject()} returns a list which includes a SingleCellExperiment object with counts, at
least one clusterization, gene markers for each clusterization, reduced dimensions for
visualization, and any additional data provided. This list is used as input when launching
the scXplorer app.
}
