% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess.R
\name{createSCEobject}
\alias{createSCEobject}
\title{Creates an object ready for use in scX}
\usage{
createSCEobject(
  xx,
  assay.name.raw = "counts",
  assay.name.normalization = "logcounts",
  metadata = NULL,
  partitionVars = NULL,
  metadataVars = NULL,
  chosen.hvg = NULL,
  nHVGs = 3000,
  nPCs = 50,
  calcRedDim = TRUE,
  paramFindMarkers = list(test.type = "wilcox", pval.type = "all", direction = "up"),
  calcAllPartitions = FALSE,
  cells2keep = NULL,
  descriptionText = NULL,
  verbose = TRUE
)
}
\arguments{
\item{xx}{Either a numeric count matrix object (with genes as rows and cells as columns), a \linkS4class{SingleCellExperiment} object or a \linkS4class{Seurat} object.}

\item{assay.name.raw}{Assay name for the raw counts matrix, if object is a \linkS4class{SingleCellExperiment}. Defaults to \code{counts}.}

\item{assay.name.normalization}{Assay name for normalized matrix if present in \linkS4class{SingleCellExperiment}. If not present,
it computes \code{logcounts} (default).}

\item{metadata}{(Optional) \linkS4class{DataFrame} with cell metadata. Rownames of the metadata data frame must include all cell names appearing as columns in \code{xx}.}

\item{partitionVars}{(Optional) Names of metadata or \code{colData} columns used for gene markers and differential
expression analysis. If \code{NULL} (default),    a quick clusterization, using \code{\link[scran]{quickCluster}} from \pkg{scran} package will be computed.}

\item{metadataVars}{(Optional) Names of additional metadata or \code{colData} columns to use for coloring in plots.
If \code{NULL} (default), only \code{partitionVars} columns will be available for coloring plots.}

\item{chosen.hvg}{(Optional) list of Highly Variable Genes. NOTE: if \code{chosen.hvg=NULL} and \code{xx} is a \linkS4class{Seurat} object with computed \code{\link[Seurat]{VariableFeatures}} then
this parameter will be set to that list of genes.}

\item{nHVGs}{Number of Highly Variable Genes to use if \code{chosen.hvg=NULL}. Defaults to 3000.}

\item{nPCs}{Number of Principal Components to use in PCA. Defaults to 50.}

\item{calcRedDim}{Logical. Indicates whether to compute reduced dimensions (PCA, UMAP, TSNE, UMAP2D,
TSNE2D). Defaults to \code{TRUE}. (Note: If it is set to \code{FALSE} but there is no 2D reduced dimension provided in the input, 2-dim PCA, UMAP and TSNE embeddings will be estimated. The same happens for 3D embeddings).}

\item{paramFindMarkers}{named List of parameters to pass to \code{\link[scran]{findMarkers}} to compute cluster gene markers.
Defaults to \code{list(test.type="wilcox", pval.type="all", direction="any")}.}

\item{calcAllPartitions}{Logical. Defaults to \code{FALSE}, meaning that only partitions from \code{partitionVars} with less or equal to 30 levels will be considered for
markers and DEGs calculations. If TRUE, force the computation of markers and DEGs for
the entire list of \code{partitionVars}.}

\item{cells2keep}{(Optional) List of cell names to keep in case of subsampling. NOTE: Subsampling is only
activated for visual purposes for large datasets, it is not used for computations. 50k cells will be used for visualization in the app.
Their indexes are stored in the \code{CELLS2KEEP} element of the CSEO object.}

\item{descriptionText}{(Optional) Short description of the object being analized. This text will be displayed in the Summary module of the \code{scX} app.}

\item{verbose}{Logical for step by step status while function is running. Defaults to \code{TRUE}.}
}
\value{
A named \linkS4class{List} that serves as input for \code{\link{launch_scX}} which contains the fields:
\describe{
\item{\code{SCE}:}{\linkS4class{SingleCellExperiment} object with a computed normalized expression,
dimensional reduction embedding (PCA, UMAP, TSNE, in 2D & 3D) calculated using the list of \code{chosen.hvg} if not \code{NULL} or
the top \code{nHVGs}. \code{colData} contains the \code{partitionVArs} and \code{metadataVars} if they where specified, if not only
a quick clusterization will be available for preliminar analysis of the data.}
\item{\code{sce.degs}:}{A named \linkS4class{List} of \linkS4class{DataFrame}s where each DataFrame contains the consolidated marker statistics
for each gene (row) for the cluster of the same name. Statistics are computed using \code{\link[scran]{findMarkers}} and user can choose \code{test.type}
parameters to pass to that function. See \code{\link[scran]{combineMarkers}} for the details of how these dataframes are confectioned.}
\item{\code{sce.markers}:}{A \linkS4class{List} of named \linkS4class{List}s of \linkS4class{DataFrame}s. Each one corresponds to the marker genes of every cluster in a partition
(names of the nested lists). \code{summary.stats:} AUC if \code{test.type=="wilcox"} and -log.FC for \code{test.type=="t"} or \code{test.type=="binom"}.
\code{log.FDR:} -Log.FDR of the most appropriate inter-cluster comparison according to the selected p-value type. See \code{\link[scran]{findMarkers}} for the details of how these metrics are computed.
\code{boxcor:} Correlation scores between the normalized gene expression profiles and a binary vector of cells, in which cells of the selected cluster have a value of 1.}
\item{(optional) \code{text}:}{String if \code{descriptionText} not \code{NULL}.}
\item{\code{CELLS2KEEP}:}{Numeric, the indices of the selected cells chosen for visualization in the \code{scX} application.
(Alternatively) Character, if 'all' no subsampling will be performed for visualization purposes.}
}
}
\description{
\code{\link{createSCEobject}} creates the input object (a \linkS4class{List}) for the function \code{\link{launch_scX}}. This list includes a \linkS4class{SingleCellExperiment} object with normalized expression,
reduced dimensions for visualization, and any additional data provided. Also, this object includes identified gene markers and differential expression analysis for each user-specified partition (if no partition was selected, a quick clusterization is computed and considered)
}
\details{
This function handles the basic preprocessing steps for sc/sn-RNAseq data. It leverages on functionality implemented in \pkg{scran}, \pkg{scater} and \pkg{SingleCellExperiment} packages.
The steps are: Converting the input to \linkS4class{SingleCellExperiment} object, the estimation of QC metrics, the normalization of the expression matrix, the identification of highly variable genes,
the computation of embeddings for various dimensionality reduction techniques. In case no cell partition is provided, a default clustering step is conducted in order to investigate marker genes and Differential expression patterns.
For large datasets (number of cells > 50k) a subsampling of the \linkS4class{SingleCellExperiment} object is considered to reduce waiting times in the \pkg{scX} app. A thorough discussion of the employed preprocessing pipeline can be found at
\href{https://bioconductor.org/books/release/OSCA/.}{OSCA} book.
}
\section{Partitions}{

It is recommended to include a curated partition of the data in the input object or in the \code{metadata}.
Marker genes and differential expression analysis automatically will be computed for partitions having less than 31 levels, that are specified as \code{partitionVars}.
If user wants to run the calculations for partitions presenting more than 30 levels \code{calcAllPartitions} must be set to \code{TRUE} (Note that this step could be very time consuming).
It may be worth coloring some partitions in plots but without having to compute marker genes and DEGs analyses for them. Those partitions
must be passed through \code{metadataVars}.
If \code{partitionVars} and \code{metadataVars} are both \code{NULL} a quickclustering step (see "Normalization") is used in order to identify markers and DEGs.
}

\section{Normalization}{

If there is no normalized assay in the \linkS4class{SingleCellExperiment} object or \linkS4class{Seurat} input object
a \href{https://bioconductor.org/books/3.17/OSCA.basic/normalization.html#normalization-by-deconvolution}{Normalization by deconvolution} is performed as proposed in the OSCA book.
First, we calculate clusters using the Walktrap community detection algorithm for graph-based clustering (default parameters from \code{\link[scran]{quickCluster}}).
The resulting clusters are stored in \code{colData} as \code{"scx.clust"}.
Then we compute scale factors for the cells using those clusters.
Finally, we calculate the lognormalized expression matrix by applying a log2 transformation to the product of the raw matrix and scale factors considering a pseudocounts of 1.
If a normalized assay exists and "scx.clust" is included in the \code{partitionVars}, the function described before will be applied to compute the clusters, which are stored in \code{colData}.
}

\section{Highly Variable Genes}{

If \code{chosen.hvg} is not specified, we will use \code{\link[scran]{modelGeneVar}} to calculate the variance and mean of the lognormalized expression values.
By fitting a trend of the variance against the mean, a biological component of variation for each gene can be assigned as the residual from the trend (see \pkg{scran} documentation for more details).
We consider the top \code{nHVGs} most variable genes with biological component > 0.
}

\section{Reduced Dimensional Analysis Techniques}{

\pkg{scX} is a tool that helps visualizing the data properly, so it runs some dimensionality reduction analysis techniques (PCA, UMAP2D, TSNE2D, UMAP3D, TSNE3D).
If \code{xx} already has dimensionality reduction embeddings calculated, \code{calcRedDim} can be set to \code{FALSE} but:
\itemize{
\item if there is no dimRed calculated in the \linkS4class{SingleCellExperiment} object -> it calculates all
\item if there is no 2D or 3D redDim caclulated -> it calculates all
\item if there is no 2D -> it calculates only 2D
\item if there is no 3D -> it calculates only 3D
}

Principal Component Analysis is calculated with \code{\link[scater]{runPCA}} using the \code{chosen.hvg} and retaining the first \code{nPCs} components,
for the normalized expression matrix.
TSNE and UMAP are calculated with \code{\link[scater]{runTSNE}} and \code{\link[scater]{runUMAP}} functions using the PCA matrix.
}

\section{Marker Genes Analysis}{

\link{createSCEobject} computes statistics to identify marker genes for every cluster
for all partitions in \code{partitionVars}
using \code{\link[scran]{findMarkers}} functions (with parameters specified by the user in \code{paramFindMarkers}).
Only genes with FDR<0.05 are selected for each cluster. The boxcor score is calculated for those genes as follows:
\describe{
\item{\code{boxcor}:}{The boxcor is the correlation between a gene's expression vector (logcounts) and a binary vector, where only the cells from the selected cluster
mark 1 while the rest of the cells mark 0.}
}
}

\section{Differential Expression Analysis}{

We use the \code{\link[scran]{findMarkers}} function to identify DEGs between clusters specified in \code{partitionVars}
(\code{direction="any"},\code{pval.type="all"}).
The test.type can be specified by the user in \code{paramFindMarkers$test.type}.
}

\section{Subsampling cells}{

If the \linkS4class{SingleCellExperiment} object contains over 50k cells, a random sample of 50k cells will be chosen for visualization purposes in the application.
Cell names that the user wants to keep in the visualizations can be specified in the \code{cells2keep} parameter.
Please note that it is solely for producing efficient visualizations.
}

\examples{
\dontrun{
# Quick start guide example:
library(scX)
cseo <- createSCEobject(xx = sce, 
                        partitionVars = "inferred_cell_type", 
                        metadataVars = c("source_name", "age", "sex", "strain", "treatment", "pseudotime"),
                        descriptionText = "Quick Start Guide")
launch_scX(cseo)
}
}
\seealso{
Related functions from \pkg{scran} and \pkg{scater} packages as suggested in \href{https://bioconductor.org/books/release/OSCA/.}{OSCA} book for:
\itemize{
\item{Preprocessing steps: }{\code{\link[scran]{quickCluster}}, \code{\link[scran]{computeSumFactors}} and \code{\link[scater]{logNormCounts}}.}
\item{HVGs: }{\code{\link[scran]{modelGeneVar}}.}
\item{Dimensionality Reduction Techniques: }{\code{\link[scater]{runPCA}}, \code{\link[scater]{runTSNE}} and \code{\link[scater]{runUMAP}}.}
\item{Marker genes and DEGs analyses: }{\code{\link[scran]{findMarkers}}, \code{\link[scran]{combineMarkers}}.}
}
}
\author{
Tomás Vega Waichman, Maria Luz Vercesi, Ariel A. Berardino, Maximiliano S. Beckel, Chernomoretz Lab and Collaborators
}
